Adding command args lenght validation

+++ FileSystem/Commands/RemoveDirCommand.cs
@@ -6,6 +6,9 @@ namespace FileSystem.Commands
     {
         public override void Execute(IFileSystemEngine fsEngine, string[] args)
         {
+            if(args.Length < 2)
+                throw new ArgumentException("The command removedir requires a path as an argument.");
+
             fsEngine.RemoveDir(args[1]);
         }
     }
